Move the `OutputFormatter` class to a separate file
This resolves a PSR-12 violation.
